Comment is now saved before subscription.
This matters because if the subscription is created first, a
"new comment" email will be sent as soon as the comment is saved.
Since the user just left that comment, that is undesireable.

Also, the general comment/subscription creation method was cleaned
up by putting the save methods within a transaction.

This fixes #2